Add files via upload
in that i made a calculator. a basic calculator in that we can perform some basic operation like addition, substraction, multiplication, etc also i added the history tab that store the operation that we are going to perform also i add in the tab is clear history button that clear the hole history of operations.